GradeOS-Platform/backend/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
alembic==1.17.2
    # via ai-grading-agent (pyproject.toml)
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.0
    # via
    #   google-genai
    #   httpx
    #   starlette
cachetools==6.2.4
    # via google-auth
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via uvicorn
colorama==0.4.6
    # via
    #   click
    #   tqdm
distro==1.9.0
    # via google-genai
fastapi==0.124.4
    # via ai-grading-agent (pyproject.toml)
filetype==1.2.0
    # via langchain-google-genai
google-ai-generativelanguage==0.6.15
    # via google-generativeai
google-api-core==2.28.1
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.187.0
    # via google-generativeai
google-auth==2.45.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-genai
    #   google-generativeai
google-auth-httplib2==0.3.0
    # via google-api-python-client
google-genai==1.55.0
    # via langchain-google-genai
google-generativeai==0.8.6
    # via ai-grading-agent (pyproject.toml)
googleapis-common-protos==1.72.0
    # via
    #   google-api-core
    #   grpcio-status
greenlet==3.3.0
    # via sqlalchemy
grpcio==1.76.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.71.2
    # via google-api-core
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httplib2==0.31.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.28.1
    # via
    #   google-genai
    #   langgraph-sdk
    #   langsmith
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
imagehash==4.3.2
    # via ai-grading-agent (pyproject.toml)
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain-core==1.2.2
    # via
    #   langchain-google-genai
    #   langgraph
    #   langgraph-checkpoint
    #   langgraph-prebuilt
langchain-google-genai==4.1.1
    # via ai-grading-agent (pyproject.toml)
langgraph==1.0.5
    # via ai-grading-agent (pyproject.toml)
langgraph-checkpoint==3.0.1
    # via
    #   langgraph
    #   langgraph-checkpoint-postgres
    #   langgraph-prebuilt
langgraph-checkpoint-postgres==3.0.2
    # via ai-grading-agent (pyproject.toml)
langgraph-prebuilt==1.0.5
    # via langgraph
langgraph-sdk==0.3.0
    # via langgraph
langsmith==0.5.0
    # via langchain-core
mako==1.3.10
    # via alembic
markupsafe==3.0.3
    # via mako
nexus-rpc==1.2.0
    # via temporalio
numpy==2.3.5
    # via
    #   imagehash
    #   pywavelets
    #   scipy
orjson==3.11.5
    # via
    #   langgraph-checkpoint-postgres
    #   langgraph-sdk
    #   langsmith
ormsgpack==1.12.1
    # via langgraph-checkpoint
packaging==25.0
    # via
    #   langchain-core
    #   langsmith
pdf2image==1.17.0
    # via ai-grading-agent (pyproject.toml)
pillow==12.0.0
    # via
    #   ai-grading-agent (pyproject.toml)
    #   imagehash
    #   pdf2image
proto-plus==1.27.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==5.29.5
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
    #   temporalio
psycopg==3.3.2
    # via
    #   ai-grading-agent (pyproject.toml)
    #   langgraph-checkpoint-postgres
psycopg-pool==3.3.0
    # via
    #   langgraph-checkpoint-postgres
    #   psycopg
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.12.5
    # via
    #   ai-grading-agent (pyproject.toml)
    #   fastapi
    #   google-genai
    #   google-generativeai
    #   langchain-core
    #   langchain-google-genai
    #   langgraph
    #   langsmith
pydantic-core==2.41.5
    # via pydantic
pymupdf==1.26.7
    # via ai-grading-agent (pyproject.toml)
pyparsing==3.2.5
    # via httplib2
python-dotenv==1.2.1
    # via ai-grading-agent (pyproject.toml)
python-multipart==0.0.21
    # via ai-grading-agent (pyproject.toml)
pywavelets==1.9.0
    # via imagehash
pyyaml==6.0.3
    # via langchain-core
redis==7.1.0
    # via ai-grading-agent (pyproject.toml)
requests==2.32.5
    # via
    #   google-api-core
    #   google-auth
    #   google-genai
    #   langsmith
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via langsmith
rsa==4.9.1
    # via google-auth
scipy==1.16.3
    # via imagehash
sniffio==1.3.1
    # via google-genai
sqlalchemy==2.0.45
    # via alembic
starlette==0.50.0
    # via fastapi
temporalio==1.20.0
    # via ai-grading-agent (pyproject.toml)
tenacity==9.1.2
    # via
    #   google-genai
    #   langchain-core
tqdm==4.67.1
    # via google-generativeai
types-protobuf==6.32.1.20251210
    # via temporalio
typing-extensions==4.15.0
    # via
    #   alembic
    #   fastapi
    #   google-genai
    #   google-generativeai
    #   grpcio
    #   langchain-core
    #   nexus-rpc
    #   psycopg-pool
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   temporalio
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.3
    # via psycopg
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.6.2
    # via requests
uuid-utils==0.12.0
    # via
    #   langchain-core
    #   langsmith
uvicorn==0.38.0
    # via ai-grading-agent (pyproject.toml)
websockets==15.0.1
    # via google-genai
xxhash==3.6.0
    # via langgraph
zstandard==0.25.0
    # via langsmith
asyncpg==0.31.0
